+++ readme.md
@@ -384,7 +384,7 @@ Thanks for contributing! 🦋🙌
 - [](# "linkify-full-profile-readme-title") [Linkifies the readme text on profile pages.](https://user-images.githubusercontent.com/16872793/90910173-ebe4bf00-e3a4-11ea-8fc5-aea3d1a2e5e5.png)
 - [](# "convert-pr-to-draft-improvements") [Moves the "Convert PR to Draft" button to the mergeability box and adds visual feedback to its confirm button.](https://user-images.githubusercontent.com/1402241/95644892-885f3f80-0a7f-11eb-8428-8e0fb0c8dfa5.gif)
 - [](# "clean-header-search-field") [Clears duplicate queries in the header search field.](https://user-images.githubusercontent.com/1402241/114177338-7c890300-9966-11eb-83a3-a711a76fae99.png)
-- [](# "cancel-edit-button") [Adds a cancel button when editing wiki pages.](https://user-images.githubusercontent.com/44045911/121776611-6abf1880-cbc0-11eb-9463-b805a5aa81f8.png)
+- [](# "cancel-wiki-edit-button") [Adds a cancel button when editing wiki pages.](https://user-images.githubusercontent.com/44045911/121776611-6abf1880-cbc0-11eb-9463-b805a5aa81f8.png)
 
 <!-- Refer to style guide above. Keep this message between sections. -->
