+++ source/refined-github.ts
@@ -221,7 +221,7 @@ import './features/avoid-accidental-submissions';
 import './features/delete-review-comments-faster';
 import './features/no-useless-split-diff-view';
 import './features/list-pr-for-branch';
-import './features/cancel-edit-button';
+import './features/cancel-wiki-edit-button';
 
 // Add global for easier debugging
 (window as any).select = select;
